requirements.txt
ansicolors
tabulate
watchdog